Fix ReadClipboardText fragment for Android 10 and above
This adds a short delay before updating the fragment's text input
layout with clipboard text on Android 10 and above. This shouldn't
be necessary, but it seems to require this delay before the
foreground application is permitted to read the clipboard.

This mainly applies to the 'Edit & read clipboard' launcher
shortcut.