Change the application's task interruption  behaviour
The application no longer interrupts running TTS tasks.  A warning
message now appears instead.  If there is a file synthesis task in
progress and the user tries to play text, the current task must fir-
st be stopped.  This also applies the other way round.

Selection menus in the settings will only play sample text if no TTS
task is in progress.  Attempts by the user to set preferences that
require TTS reinitialisation will fail if a TTS task is in progress.