Add playback-on-input feature
Re: #21.

This feature automatically reads changes to the application's input
text fields.  Each word is spelt out as it is typed and spoken toge-
ther when finished, i.e., when a space or other delimiter is typed.
If the entire text field is replaced, then it is read from beginning
to end normally.

This feature is disabled by default and may be enabled by the option
in the settings screen.