Fix the handling of empty/blank input sources
When the input fields contain blank (whitespace-only) text, it is now
"read" and synthesized correctly.  It now makes sense to do this given
given the setting for custom silence in place of line endings.

The warning about zero-length input is now suppressed for the *Read
text* and *Read clipboard* screens when the 'Play' button is pressed.