Improve the behaviour of the reinitialise text-to-speech button
The main activity's reinitialise text-to-speech button now shows a
confirmation dialog if pressed when a task is in progress.  It does
not do this if sample text is in progress.  This exception makes
sense to me currently, but I might change it in the future.

Two resource strings have been added for the alert dialog.